pyproject.toml
[tool.poetry]
name = "vertin-admin"
version = "0.1.0"
description = "Vertin Admin"
authors = ["fireinrain <fireinrain@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
tortoise-orm = "^0.19.3"
pydantic = "^2.3.0"
email-validator = "^2.0.0.post2"
passlib = "^1.7.4"
pyjwt = "^2.7.0"
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.281"
loguru = "^0.7.0"
pydantic-settings = "^2.0.3"
argon2-cffi = "^23.1.0"

[tool.black]
line-length = 120
target-version = ["py310", "py311"]

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[tool.ruff]
line-length = 120
extend-select = [
  # "I",    # isort
#   "B",    # flake8-bugbear
#   "C4",   # flake8-comprehensions
#   "PGH",  # pygrep-hooks
  # "RUF",  # ruff
#   "W",    # pycodestyle
#   "YTT",  # flake8-2020
]
ignore = [
    "F403",
    "F405",
]